requirements.txt
requests==2.23.0
botbuilder-core==4.13.0
botframework-connector==4.13.0
botbuilder-ai==4.13.0 
botbuilder-schema==4.13.0
botbuilder-dialogs==4.13.0
botbuilder-applicationinsights==4.13.0
azure-cognitiveservices-language-luis==0.2.0
botbuilder-integration-applicationinsights-aiohttp==4.13.0
msrest==0.6.10
datatypes-date-time>=1.0.0.a1
aiohttp==3.6.2
pytest